requirements.txt
beautifulsoup4==4.9.3
decorator==5.0.9
Flask==0.10.1
Flask-Assets==0.10
Flask-Classy==0.6.10
Flask-KVSession==0.6.1
Flask-Login==0.2.11
Flask-ReCaptcha==0.3
Flask-S3==0.1.7
gitdb==4.0.7
GitPython==3.1.18
httplib2==0.19.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lxml==4.6.3
matplotlib==3.3.4
mysql-connector-python==8.0.25
notebook==6.4.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
plotly==5.1.0
PyMySQL==0.6.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
sklearn==0.0
smmap==4.0.0
SQLAlchemy==0.9.8
streamlit==0.83.0
Unidecode==1.2.0
urllib3==1.26.6
wordcloud==1.8.1
zipp==3.4.1
